v2.1
In this update, a minor stylistic modification was made to ensure font consistency across all elements of the webpage, particularly focusing on the text inside certain buttons. Previously, the text within the buttons "IP تصادفی", "تبدیل", "تحلیل", "آزمون", and "ریست کردن" was displayed in a different font due to the application of the Material Icons font from the material-icons class on the span elements within each button.

To address this, a specific CSS rule was added to the style section of the HTML document, enforcing the 'Vazirmatn' font to be used on the text inside all button elements. This rule was placed within the style section of the HTML, ensuring it overrides any conflicting font-family settings, thereby maintaining a consistent font appearance across the entire webpage.

This change enhances the visual consistency and readability of the text, providing a better user experience.